Move attach and highlight state out of proof state
