--- homotopy-model/tests/examples/braid_contraction.json
@@ -1,251 +1,0 @@
-[true,
-[
-  "CreateGeneratorZero",
-  {
-    "SelectGenerator": {
-      "id": 0,
-      "dimension": 0
-    }
-  },
-  {
-    "SetBoundary": "Source"
-  },
-  {
-    "SelectGenerator": {
-      "id": 0,
-      "dimension": 0
-    }
-  },
-  {
-    "SetBoundary": "Target"
-  },
-  {
-    "SelectGenerator": {
-      "id": 1,
-      "dimension": 1
-    }
-  },
-  {
-    "EditSignature": {
-      "Remove": 2
-    }
-  },
-  {
-    "SelectGenerator": {
-      "id": 0,
-      "dimension": 0
-    }
-  },
-  "TakeIdentityDiagram",
-  {
-    "SetBoundary": "Source"
-  },
-  {
-    "SelectGenerator": {
-      "id": 0,
-      "dimension": 0
-    }
-  },
-  "TakeIdentityDiagram",
-  {
-    "SetBoundary": "Target"
-  },
-  {
-    "SelectGenerator": {
-      "id": 1,
-      "dimension": 2
-    }
-  },
-  {
-    "SetBoundary": "Target"
-  },
-  {
-    "SelectGenerator": {
-      "id": 0,
-      "dimension": 0
-    }
-  },
-  "TakeIdentityDiagram",
-  "TakeIdentityDiagram",
-  {
-    "SetBoundary": "Source"
-  },
-  {
-    "SelectGenerator": {
-      "id": 2,
-      "dimension": 3
-    }
-  },
-  "ClearWorkspace",
-  {
-    "SelectGenerator": {
-      "id": 1,
-      "dimension": 2
-    }
-  },
-  "TakeIdentityDiagram",
-  {
-    "AscendSlice": 1
-  },
-  {
-    "SelectPoints": [
-      [
-        "Target",
-        "Target"
-      ],
-      [
-        {
-          "Regular": 0
-        },
-        {
-          "Regular": 1
-        }
-      ],
-      [
-        {
-          "Regular": 0
-        },
-        "Target"
-      ]
-    ]
-  },
-  {
-    "Homotopy": {
-      "Contract": {
-        "bias": "Lower",
-        "location": [
-          "Target"
-        ],
-        "height": 0,
-        "direction": "Forward"
-      }
-    }
-  },
-  {
-    "Homotopy": {
-      "Expand": {
-        "location": [
-          "Target",
-          {
-            "Singular": 0
-          },
-          {
-            "Singular": 0
-          }
-        ],
-        "direction": "Forward"
-      }
-    }
-  },
-  {
-    "Homotopy": {
-      "Contract": {
-        "bias": "Lower",
-        "location": [],
-        "height": 1,
-        "direction": "Backward"
-      }
-    }
-  },
-  {
-    "SelectPoints": [
-      [
-        "Target",
-        {
-          "Regular": 1
-        }
-      ]
-    ]
-  },
-  {
-    "SelectPoints": [
-      [
-        "Target",
-        {
-          "Regular": 2
-        }
-      ],
-      [
-        "Target",
-        {
-          "Singular": 2
-        }
-      ],
-      [
-        {
-          "Regular": 2
-        },
-        {
-          "Singular": 2
-        }
-      ]
-    ]
-  },
-  {
-    "SelectPoints": [
-      [
-        "Target",
-        {
-          "Regular": 3
-        }
-      ],
-      [
-        {
-          "Regular": 3
-        },
-        {
-          "Singular": 2
-        }
-      ],
-      [
-        {
-          "Regular": 3
-        },
-        {
-          "Regular": 3
-        }
-      ]
-    ]
-  },
-  {
-    "SelectPoints": [
-      [
-        "Target",
-        {
-          "Regular": 4
-        }
-      ]
-    ]
-  },
-  {
-    "Homotopy": {
-      "Contract": {
-        "bias": "Lower",
-        "location": [],
-        "height": 2,
-        "direction": "Backward"
-      }
-    }
-  },
-  {
-    "Homotopy": {
-      "Contract": {
-        "bias": "Lower",
-        "location": [],
-        "height": 2,
-        "direction": "Backward"
-      }
-    }
-  },
-  {
-    "Homotopy": {
-      "Contract": {
-        "bias": "Lower",
-        "location": [],
-        "height": 2,
-        "direction": "Backward"
-      }
-    }
-  }
-]
-]
